pyproject.toml
[tool.poetry]
name = "django-todo"
version = "0.1.0"
description = ""
authors = ["mauricio <mauricioduhamsolis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.6"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
djoser = "2.2.2"
drf-spectacular = "0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
